# laravel-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ChefMate",
        "description": "API para gestión de usuarios, recetas, listas y comentarios de ChefMate.",
        "contact": {
            "name": "Francisco Javier Cabello Rueda"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor local de desarrollo"
        },
        {
            "url": "http://chefmate-internal.duckdns.org/api",
            "description": "Servidor de producción"
        }
    ],
    "paths": {
...
        "/listas": {
            "get": {
                "tags": [
                    "Listas"
                ],
                "summary": "Listar listas del usuario autenticado",
                "operationId": "71f6ad3cb23f11888678cb333c9950b7",
                "responses": {
                    "200": {
                        "description": "Listas obtenidas correctamente"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Listas"
                ],
                "summary": "Crear una nueva lista",
                "operationId": "9f3529d9b2d5eee6e1bcf8715c9f57d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_lista"
                                ],
                                "properties": {
                                    "nombre_lista": {
                                        "type": "string",
                                        "example": "Recetas favoritas"
                                    },
                                    "recetas": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lista creada correctamente"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/listas/{id}": {
            "put": {
                "tags": [
                    "Listas"
                ],
                "summary": "Actualizar una lista",
                "operationId": "3eaaea5472d0b483ba1894fb44e9a9dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la lista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_lista": {
                                        "type": "string",
                                        "example": "Actualizado"
                                    },
                                    "recetas": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lista actualizada correctamente"
                    },
                    "404": {
                        "description": "Lista no encontrada"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Listas"
                ],
                "summary": "Eliminar una lista",
                "operationId": "161c125eea6d7ec9fa79833b73b44e50",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la lista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista eliminada correctamente"
                    },
                    "404": {
                        "description": "Lista no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/recetas": {
            "get": {
                "tags": [
                    "Recetas"
                ],
                "summary": "Listar todas las recetas con su autor",
                "operationId": "c255512219e7a394296629ad4d5fced2",
                "responses": {
                    "200": {
                        "description": "Recetas obtenidas correctamente"
                    },
                    "404": {
                        "description": "No se encontraron recetas"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/recetas/{id}": {
            "get": {
                "tags": [
                    "Recetas"
                ],
                "summary": "Obtener una receta específica con sus detalles",
                "operationId": "ba970950c4980e1e16985216873967e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la receta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receta obtenida correctamente"
                    },
                    "404": {
                        "description": "Receta no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de acceso obtenido en /login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}